feat: anecdotes to test
